Fix coding errors.
For Sonic's 32nd birthday, a bunch of things have been fixed, bringing this project one step closer to actually running.

This involves optimizing and editing a good chunk of the code to do some of the non-CPU routines of the Master System in ASM, from the address pointer to paging in ROM banks.

A playable build will take quite a while to emerge, but until then, thank you for following this project.